% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_counts.R
\name{addCardDeck}
\alias{addCardDeck}
\title{Add a deck}
\usage{
addCardDeck(deck_name)
}
\arguments{
\item{deck_name}{Name of the deck}
}
\description{
Downloads a card deck from www.dontyouremember.com ( see TODOTODO for a list
of all card decks )
}
\examples{
addCardDeck( "advanced-r-object-oriented-part-1" )

}
