% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_initial_db.R
\name{create_initial_db}
\alias{create_initial_db}
\title{When a new user first launches remembr, we want them to first learn how to use remembr!}
\usage{
create_initial_db()
}
\description{
When a new user first launches remembr, we want them to first learn how to use remembr!
}
