% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packs.R
\name{createCardDeck}
\alias{createCardDeck}
\title{Create a deck of cards from files}
\usage{
createCardDeck(name, files)
}
\description{
Creates a new card deck and stores it in a local folder called 'packs'
(this is an alternative method (  and perhaps better ) than the method above!
}
\details{
@param name name of the deck
@files a vector of file names

@export
}
