% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_counts.R
\name{.convertEnvToDataFrame}
\alias{.convertEnvToDataFrame}
\title{helper function for convertCallCountsToHashTable}
\usage{
.convertEnvToDataFrame(call_counts_hash_table)
}
\arguments{
\item{call_counts_hash_table}{A Call counts hash table ( like from
getCallCountsHashTable )}
}
\description{
helper function for convertCallCountsToHashTable
}
