% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl_control.R
\name{echoExpressionCallback}
\alias{echoExpressionCallback}
\title{This just echos the input expression.  It is here for testing out the
taskCallback code}
\usage{
echoExpressionCallback(expr, value, status, visible, data)
}
\arguments{
\item{expr}{s-language expression}

\item{value}{result of the expression evaluation}

\item{status}{logical indicating success or not}

\item{visible}{was the output printed}

\item{data}{data object that is accessible to the callback ( passed in from addTaskCallback)}
}
\value{
I guess this indicates if the callback succeeded?
}
\description{
This just echos the input expression.  It is here for testing out the
taskCallback code
}
