% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packs.R
\name{finishPack}
\alias{finishPack}
\title{Finish Pack}
\usage{
finishPack(persist = TRUE)
}
\arguments{
\item{persist}{Do you want to save the pack to your file system? TRUE or FALSE}
}
\description{
Allows for creation of a "pack" of flashcards.  The idea is that you
startPack; then code as normal; and then endPack.  Any code entered in between
will become part of a custom card pack
}
