% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_url.R
\name{getDocumentationURL}
\alias{getDocumentationURL}
\title{Get documentation for a url}
\usage{
getDocumentationURL(targetFunction)
}
\description{
Pass in a function and get any urls associated with it
}
