% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_url.R
\name{getDocumentationURLs}
\alias{getDocumentationURLs}
\title{Get documentation for a function}
\usage{
getDocumentationURLs(targetFunction, call_counts_hash_table)
}
\description{
Pass in a function and get any urls associated with it
}
