% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_counts.R
\name{getDurationFromBucketId}
\alias{getDurationFromBucketId}
\title{Given a bucket id, return a timer for that bucket}
\usage{
getDurationFromBucketId(bucket_id)
}
\description{
Given a bucket id, return a timer for that bucket
}
