% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getFunctionsFromFile}
\alias{getFunctionsFromFile}
\title{Get functions from a file}
\usage{
getFunctionsFromFile(paths)
}
\description{
Get functions from a file
}
\examples{
env = getFunctionsFromFile('~/git/leitnr/R/get_functions.R' )

}
