% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{initRemembr}
\alias{initRemembr}
\title{Initialize remembr}
\usage{
initRemembr()
}
\description{
This tells R to record your R code, as your are coding, in order to build
your personal flashcards and list of function calls.  It will remain on for
the duration of your R session.
This adds a callback handler to your REPL.
It will track what you do in a local file; and only track function and
package names.
}
