% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{install_remembr}
\alias{install_remembr}
\title{Adds a line to your .Rprofile so that remembr runs by default.
Can be uninstalled at any time with uninstall_remembr}
\usage{
install_remembr()
}
\description{
Adds a line to your .Rprofile so that remembr runs by default.
Can be uninstalled at any time with uninstall_remembr
}
