% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_counts.R
\name{nextBucket}
\alias{nextBucket}
\title{Next bucket}
\usage{
nextBucket(bucket_id)
}
\arguments{
\item{bucket_id}{the id of the current bucket}
}
\description{
Given an id, get the id for the next bucket
If bucket_id is null, then start from 1
}
