% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remembr.R
\docType{package}
\name{remembr}
\alias{remembr}
\alias{remembr-package}
\title{remembr: Learn and remember R}
\description{
To install the package just use remembr::installRemembr.  This will add a line to your .rProfile, which you can remove at any time.
You can learn more here www.dontyouremember.com
}
\section{Flashcards Just call the flashcards method to review what you have learned}{
NA
}

