% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reminders.R
\name{remindMe}
\alias{remindMe}
\title{Remind Me}
\usage{
remindMe(num_rows = 5)
}
\arguments{
\item{num_rows}{how many functions to show}
}
\description{
Shows a list of functions that are upcoming for you to study.  This is based
on your personal R history, and any flashcard decks you may use.
}
