% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reminders.R
\name{remindPackage}
\alias{remindPackage}
\title{Shows upcoming reminders for a specific package}
\usage{
remindPackage(packageNames = "base", all = FALSE)
}
\arguments{
\item{packageNames}{a list of package names}

\item{all}{if all is true then shows all packages}
}
\value{
a data frame with the upcoming reminder.
}
\description{
Shows upcoming reminders for a specific package
}
