% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packs.R
\name{startPack}
\alias{startPack}
\title{Start Pack}
\usage{
startPack(pack_name)
}
\arguments{
\item{pack_name}{the name of the pack}
}
\description{
Allows for creation of a "pack" of flashcards.  The idea is that you
startPack; then code as normal; and then endPack.  Any code entered in between
will become part of a custom card pack
}
