% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reminders.R
\name{upcomingReminders}
\alias{upcomingReminders}
\title{Upcomping reminders}
\usage{
upcomingReminders(num_methods = 10)
}
\arguments{
\item{num_methods}{How many methods to show}
}
\description{
Prints out the methods that will need to be reviewed next time you do a review
}
