% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reminders.R
\name{convertCallCountsToHashTable}
\alias{convertCallCountsToHashTable}
\title{Helper method for parsing the call_counts_hash_table environment and presenting it as a data frame}
\usage{
convertCallCountsToHashTable(call_counts_hash_table)
}
\description{
Helper method for parsing the call_counts_hash_table environment and presenting it as a data frame
}
