% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reminders.R
\name{flashCards}
\alias{flashCards}
\title{Flashcards for Code}
\usage{
flashCards(num_flashcards = 10, time_since_last_use = NULL,
  pack_name = NULL, pack = NULL)
}
\arguments{
\item{num_flashcards}{how many flashcards to do at once.}

\item{time_since_last_use}{time interval since the last time the function was used in console.  For example: lubridate::hours(1)}
}
\description{
Shows flashcards for the methods that will help you become more fluent at R.
You will gave a short practice session to review each method.
If you say you are comfortable with the method, you will not see it for a while.
If you are not comfortable, it will return to your flashcards so you have another chance to review.
}
