% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_url.R
\name{getDocumentationURLs}
\alias{getDocumentationURLs}
\title{Get documentation for a function}
\usage{
getDocumentationURLs(targetFunction, call_counts_hash_table = NULL)
}
\arguments{
\item{targetFunction}{a function reference that you want to get documentation urls for}

\item{call_counts_hash_table}{the call counts hash table ( uses remebmr one by default )}
}
\description{
Pass in a function and get any urls associated with it
}
