% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getFunctionsFromFiles}
\alias{getFunctionsFromFiles}
\title{Get functions from a file}
\usage{
getFunctionsFromFiles(paths, output_env = NULL, libraries = NULL)
}
\description{
Get functions from a file
}
\examples{
env = getFunctionsFromFile('~/git/leitnr/R/get_functions.R' )

TODO: should output a list of missing pacakges

}
