% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_counts.R
\name{mergeCallCountHashTables}
\alias{mergeCallCountHashTables}
\title{Code for merging two hash tables together}
\usage{
mergeCallCountHashTables(call_counts_hash_table1, call_counts_hash_table2,
  in_place = FALSE)
}
\description{
Code for merging two hash tables together
}
