% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reminders.R
\name{upcomingReminders}
\alias{upcomingReminders}
\title{Prints out the methods that will need to be reviewed next time you do a review}
\usage{
upcomingReminders(num_methods = 10)
}
\description{
Prints out the methods that will need to be reviewed next time you do a review
}
